requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.2.3
email-validator==2.0.0
PyGitHub==1.58.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1